# docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: private-cloud API
  version: v1
  description: |
    The endpoints this server actually serves, generated from its route table.

    This file is deliberately narrow. It answers existence, method and auth
    posture — the things a client blocks on and the things that can be derived
    from the server itself. It does NOT describe request and response bodies:
    those live in docs/api-contract.md, which is written by hand and is the
    source of truth for what an endpoint means.

    The two documents have different jobs, and the difference matters. The
    contract also specifies endpoints that are only PROPOSED (Phases 5-9), so a
    client cannot use it to tell what is implemented. This file contains no
    proposed endpoints by construction. If it is not here, it is not served.

    Auth posture is derived by probing each route without credentials, not
    declared, so it cannot go stale when a handler's wrapper changes.
servers:
  - url: /
    description: Same origin as the web app — WebAuthn binds passkeys to an origin.
paths:
  "/api/v1/admin/fsck":
    post:
      operationId: postApiV1AdminFsck
      x-required-role: admin
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/albums":
    get:
      operationId: getApiV1Albums
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postApiV1Albums
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/albums/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getApiV1AlbumsId
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: patchApiV1AlbumsId
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    delete:
      operationId: deleteApiV1AlbumsId
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/albums/{id}/items":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    post:
      operationId: postApiV1AlbumsIdItems
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: patchApiV1AlbumsIdItems
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/albums/{id}/items/{nodeId}":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      - name: nodeId
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteApiV1AlbumsIdItemsNodeId
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/auth/app-passwords":
    get:
      operationId: getApiV1AuthAppPasswords
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: postApiV1AuthAppPasswords
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/auth/app-passwords/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteApiV1AuthAppPasswordsId
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/auth/credentials":
    get:
      operationId: getApiV1AuthCredentials
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/auth/credentials/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteApiV1AuthCredentialsId
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  "/api/v1/auth/me":
    get:
      operationId: getApiV1AuthMe
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  "/api/v1/auth/recovery/regenerate":
    post:
      operationId: postApiV1AuthRecoveryRegenerate
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  "/api/v1/auth/sessions":
    get:
      operationId: getApiV1AuthSessions
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/auth/sessions/{id}":
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    delete:
      operationId: deleteApiV1AuthSessionsId
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  "/api/v1/auth/token":
    post:
      operationId: postApiV1AuthToken
      security:
        - appPassword: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/changes":
    get:
      operationId: getApiV1Changes
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/chunks/have":
    post:
      operationId: postApiV1ChunksHave
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/chunks/{hash}":
    parameters:
      - name: hash
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getApiV1ChunksHash
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    put:
      operationId: putApiV1ChunksHash
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/folders":
    post:
      operationId: postApiV1Folders
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/manifests":
    post:
      operationId: postApiV1Manifests
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/media/timeline":
    get:
      operationId: getApiV1MediaTimeline
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/nodes/resolve":
    get:
      operationId: getApiV1NodesResolve
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  "/api/v1/nodes/root":
    get:
      operationId: getApiV1NodesRoot
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "default":
          description: |
            See docs/api-contract.md for this endpoint's request and response
            bodies. Every failure uses the Error schema below.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: pc_session
      description: Set by the WebAuthn, recovery-code and OIDC login flows.
    bearerToken:
      type: http
      scheme: bearer
      description: |
        A device token from POST /api/v1/auth/token. Used by pcsync and other
        headless clients.
    appPassword:
      type: http
      scheme: basic
      description: |
        A username and an app password, which is shown once at creation and
        stored argon2id-hashed. Accepted ONLY at POST /api/v1/auth/token and
        over /dav — an app password cannot reach the rest of the JSON API, so a
        leaked one cannot enrol a passkey or take the account.
...